Pulled UnmanagedInterfaceFilter out as a separate class and linked it into the RangeChunker. Brought all of the Camel actor beans into applicationContext-discovery.xml.
